refactored inspector as system
